used new ProtoModelConverter
